+++ Medicinska klinika 3.0/Medicinska klinika 3.0/Entiteti/Zaposlen.cs
@@ -8,7 +8,7 @@ namespace MedicinskaKlinika.Entiteti
 {
     public abstract class Zaposlen
     {
-        public virtual int JMBG { get; set; }
+        public virtual long JMBG { get; set; }
         public virtual DateTime DatumZaposlenja { get; set; }
         public virtual String Pozicija { get; set; }
