+++ Medicinska klinika 3.0/Medicinska klinika 3.0/IzmeniForme/IzmeniAdministrativnoOsoblje.cs
@@ -27,7 +27,7 @@ namespace Medicinska_klinika_3._0.FormeIzmeni
                 return;
             }
             ListViewItem selektovaniItem = listView1.SelectedItems[0];
-            int id = (int)selektovaniItem.Tag;
+            long id = (long)selektovaniItem.Tag;
 
             AdministrativnoOsoblje admin = DTOManager.nadjiAdministrativnoOsoblje(id);
 
@@ -90,7 +90,7 @@ namespace Medicinska_klinika_3._0.FormeIzmeni
                 return;
             }
             ListViewItem selektovaniItem = listView1.SelectedItems[0];
-            int id = (int)selektovaniItem.Tag;
+            long id = (long)selektovaniItem.Tag;
 
             DTOManager.brisiAdmina(id);
