+++ Medicinska klinika 3.0/Medicinska klinika 3.0/IzmeniForme/IzmeniLaboranta.cs
@@ -34,7 +34,7 @@ namespace Medicinska_klinika_3._0.FormeIzmeni
                 return;
             }
             ListViewItem selektovaniItem = listView1.SelectedItems[0];
-            int id = (int)selektovaniItem.Tag;
+            long id = (long)selektovaniItem.Tag;
 
             Laborant admin = DTOManager.nadjiLaboranta(id);
 
@@ -94,7 +94,7 @@ namespace Medicinska_klinika_3._0.FormeIzmeni
                 return;
             }
             ListViewItem selektovaniItem = listView1.SelectedItems[0];
-            int id = (int)selektovaniItem.Tag;
+            long id = (long)selektovaniItem.Tag;
 
             DTOManager.brisiLaboranta(id);
             IzmeniLaboranta_Load(sender, e);
